// src/types/hr-management/resignation-management.d.ts
export interface IResignation {
    resignationFormId: number;
    employeeId: number;
    badgeId: string;
    fullName: string;
    personalEmail: string;
    workEmail: string;
    mobilePhone: string;
    companyName: string;
    dgName: string;
    dcName: string;
    projectName: string;
    positionName: string;
    grade: string;
    beforeWorkExp: number;
    currentWorkExp: number;
    totalWorkExp: number;
    paMark: number;
    personalContact: string;
    resignationTypeId: number;
    resignationTypeName: string;
    resignationStatusId: number;
    resignationStatusName: string;
    resignationStatusColor: string;
    applyDate: string;
    resignDate: string;
    cancelledDate: string;
    reasonId: number;
    reasonName: string;
    notes: string;
    isSentResignationEmail: boolean;
    isAttrition: boolean;
    isCommitmentBound: boolean;

    birthday: string;
    birthPlace: string;
    idCardNo: string;
    idCardIssuePlace: string;
    idCardIssueDate: string;
    isOfficialEmployee: boolean;
    permanentAddress: string;

    companyOwner: string;
    companyAddress: string;
    companyPhone: string;
    companyFax: string;
}